fix: Change COUNT(*) to return Integer instead of Numeric (#1204)
* fix: Change COUNT(*) to return Integer instead of Numeric

- Modified COUNT(*) to return SqlValue::Integer in three key locations:
  - Fast path for simple COUNT(*) queries
  - COUNT(*) within aggregation context
  - AggregateAccumulator::finalize() method

- Updated all affected tests to expect Integer from COUNT(*)

- Note: Integer arithmetic remains Integer (no automatic promotion to Numeric)

Fixes #1199

* fix: Handle both Integer and Numeric in test_phase3_1_equijoin_selectivity

The test was only expecting Numeric from COUNT(*), but after fixing COUNT(*)
to return Integer (as per SQL-92 standard), this test needed to be updated
to handle both types. This follows the same pattern used in the cartesian_count
match statement earlier in the same test.

* fix: Handle both Integer and Numeric in predicate pushdown integration tests

Updated three COUNT(*) match statements to handle both Integer and Numeric types:
- test_phase_2_table_local_pushdown
- test_phase_3_equijoin_hash_join
- test_phases_2_and_3_combined

This follows the same pattern as the earlier fix in predicate_pushdown_tests.

* fix: Update test_e2e_group_by_count to expect Integer from COUNT(*)

Changed expectations from Numeric(2.0) and Numeric(1.0) to Integer(2) and
Integer(1) to match the new COUNT(*) return type.

+++ crates/vibesql-executor/src/select/executor/aggregation/evaluation/aggregate_function.rs
@@ -47,7 +47,7 @@ pub(super) fn evaluate(
                 ));
             }
             // Fast path: COUNT(*) without DISTINCT is just row count (O(1) vs O(n))
-            let result = vibesql_types::SqlValue::Numeric(group_rows.len() as f64);
+            let result = vibesql_types::SqlValue::Integer(group_rows.len() as i64);
             // Cache the result
             executor.aggregate_cache.borrow_mut().insert(cache_key, result.clone());
             return Ok(result);
